langgraph/personal_coach_agent_with_mem0/pyproject.toml
[project]
name = "personal-coach-agent"
version = "0.0.1"
description = "Mem0-powered personal coaching agent using LangGraph."
authors = [{ name = "Course Demo", email = "demo@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.6",
    "python-dotenv>=1.0.1",
    "langchain-anthropic>=0.2.0",
    "langchain-core>=0.3.0",
    "pydantic>=2.0.0",
    "langgraph-cli[inmem]>=0.4.2",
    "langchain-openai>=0.2.0",
    "httpx>=0.27.0",
    "mem0ai[graph]",
    "qdrant-client>=1.11.0",
]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent"]
[tool.setuptools.package-dir]
"agent" = "src/agent"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
extend-exclude = [".venv", "**/.venv/**"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = ["anyio>=4.7.0", "mypy>=1.13.0", "pytest>=8.3.5", "ruff>=0.8.2"]
